ignoring dot files on read and delete

+++ index.js
@@ -77,7 +77,7 @@ module.exports = function(yml_file, port, log_loc){
     var file = files.shift();
     return fs.isFile(loc+file)
       .then(function(is_file){
-        if(is_file){
+        if(is_file && !file.match(/(^|\/)\./)){
           return fs.read(loc+file)
                .then(function(data){
                  try{
@@ -94,7 +94,12 @@ module.exports = function(yml_file, port, log_loc){
 
   function deleteThese(loc, files){
     var file = files.shift();
-    return fs.remove(loc+file)
+    return fs.isFile(loc+file)
+              .then(function(){
+                if(!file.match(/(^|\/)\./)){
+                  return fs.remove(loc+file)
+                }
+              }
              .then(function(){
                return deleteThese(loc, files);
              })
